% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{get_users_incremental}
\alias{get_users_incremental}
\title{Get a list of users from Zendesk that changed from a particular start time}
\usage{
get_users_incremental(connection, start_time)
}
\arguments{
\item{connection}{An connection created by the create_connection function}

\item{start_time}{A Date or POSIXCT datetime}
}
\value{
a dataframe of users
}
\description{
Get a list of users from Zendesk that changed from a particular start time
}

