% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{get_tickets_incremental}
\alias{get_tickets_incremental}
\title{Get a list of tickets from Zendesk that changed from a particular start time}
\usage{
get_tickets_incremental(connection, start_time)
}
\arguments{
\item{connection}{An connection created by the create_connection function}

\item{start_time}{A Date or POSIXCT datetime}
}
\value{
a dataframe of tickets
}
\description{
Get a list of tickets from Zendesk that changed from a particular start time
}

