% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{get_users_incremental}
\alias{get_users_incremental}
\title{Get a list of users from Zendesk that changed from a particular start time}
\usage{
get_users_incremental(access_token, email_id, start_time)
}
\arguments{
\item{access_token}{A Zendesk access token}

\item{email_id}{Email address of admin account to query using}

\item{start_time}{A Date or POSIXCT datetime}
}
\value{
a dataframe of users
}
\description{
Get a list of users from Zendesk that changed from a particular start time
}

